% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontes.R
\name{basic_call}
\alias{basic_call}
\title{Make a call on IpeaData api.}
\usage{
basic_call(api, type = "data.table")
}
\arguments{
\item{api}{Url to be called}

\item{type}{The type of return, it could be data.table or tibble.}
}
\value{
a data.table or tibble
}
\description{
\code{basic_call} return data from ipeaData
}
\examples{
   fontes <- get_fonts()
}
