% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontes.R
\name{get_fonts}
\alias{get_fonts}
\title{Return fonts from ipeaData .}
\usage{
get_fonts(type = "data.table")
}
\arguments{
\item{type}{The type of return, it could be data.table or tibble.}
}
\value{
a data.table with all fonts on ipeaData
}
\description{
\code{get_fonts} return fonts from ipeaData
}
