% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaData.R
\name{get_metadata}
\alias{get_metadata}
\title{Return metadadata from a serie on Ipeadata .}
\usage{
get_metadata(serie = NULL, type = "data.table")
}
\arguments{
\item{serie}{The serie number.}

\item{type}{The type of return, it could be data.table or tibble.}
}
\value{
a data.table with all fonts on ipeaData
}
\description{
\code{get_metadata} Return metadadata from a serie on Ipeadata.
}
