% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaData.R
\name{get_sources}
\alias{get_sources}
\title{Return series sources from ipeaData. This functions
can be used fot search for SERCODIGO of series. SERCODIGO
must be used for get series values.}
\usage{
get_sources(type = "data.table")
}
\arguments{
\item{type}{The type of return, it could be data.table or tibble.}
}
\value{
a data.table with all fonts on ipeaData
}
\description{
\code{get_sources} return series sources from ipeaData
}
