% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontes.R
\name{get_values}
\alias{get_values}
\title{Return values from a serie.}
\usage{
get_values(serie, type = "data.table")
}
\arguments{
\item{serie}{The serie number.}

\item{type}{They type of return, it could be data.table or tibble.}
}
\value{
a data.table with all fonts on ipeaData
}
\description{
\code{get_values} return values from a serie.
}
\examples{
   fontes <- get_values()
}
