% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaData.R
\name{ipeadata}
\alias{ipeadata}
\title{Return values from a serie, its metadados and citation.}
\usage{
ipeadata(serie, type = "data.table")
}
\arguments{
\item{serie}{The serie number.}

\item{type}{They type of return, it could be data.table or tibble.}
}
\value{
a data.table or tibble with all fonts on ipeaData
}
\description{
\code{ipeadata} return values from a serie.
}
\examples{
    data <- ipeadata('ADMIS')
\donttest{
    data <- ipeadata('ABATE_ABPEBV')
    data <- ipeadata('ABRACAL_PCALCARIO')
}

}
