% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontes.R
\name{make_citation}
\alias{make_citation}
\title{Return citation for a given serie.}
\usage{
make_citation(data)
}
\arguments{
\item{data}{They type of return, it could be data.table or tibble.}
}
\value{
return a string
}
\description{
\code{make_citation} Return citation for a given serie.
}
