% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaData.R
\name{output_control}
\alias{output_control}
\title{Return data from api in data.table or tibble.}
\usage{
output_control(data, type = "data.table")
}
\arguments{
\item{data}{Data.frame to be converted}

\item{type}{They type of return, it could be data.table or tibble.}
}
\value{
a data.table or tibble
}
\description{
\code{output_control} Return data from api in data.table or tibble
}
