% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaData.R
\name{search_serie}
\alias{search_serie}
\title{Search for series base on a term.}
\usage{
search_serie(term = NULL, fields_search = c("SERCODIGO", "SERNOME",
  "PERNOME", "UNINOME"), type = "data.table")
}
\arguments{
\item{term}{The term to search for.}

\item{fields_search}{fields to looking in.}

\item{type}{The type of return, it could be data.table or tibble.}
}
\value{
a data.table with all fonts on ipeaData
}
\description{
\code{search_serie} Return series' data from a query on Ipeadata.
}
