% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaData.R
\name{format_data}
\alias{format_data}
\title{Return data formated for citation .}
\usage{
format_data(date_value)
}
\arguments{
\item{date_value}{data as Date object}
}
\description{
\code{format_data} return data formated for citation
}
